requirements.txt
requests
beautifulsoup4
pandas
gspread
oauth2client
pytest
python-dotenv
